Ignorecase on Telport Command
I am on holiday, which is why I cannot check my changes. But I have heard that it bothers some that /tpll is case sensetive. I figured it would be easy to just make the input lowercase. If that doesn't work I apologise and would be happy if someone else could fix it. 
Kind regards